test: Split checkRedirectIntervalLimit test to another file
And enable to test it for Chrome add-on.
