+++ webextensions/Makefile
@@ -8,15 +8,18 @@ all: packages
 testee:
 	mkdir $@
 
+testee/chrome.js: edge/background.js testee
+	sed -e "s/ThinBridgeTalkClient.init();/exports.client = ThinBridgeTalkClient/g" -e "s/ResourceCap.init();//g" chrome/background.js > $@
+
 testee/edge.js: edge/background.js testee
 	sed -e "s/ThinBridgeTalkClient.init();/exports.client = ThinBridgeTalkClient/g" -e "s/ResourceCap.init();//g" edge/background.js > $@
 
-unittest: install_dependency testee/edge.js
+unittest: install_dependency testee/chrome.js testee/edge.js
 	npx mocha --require mocha-suppress-logs test
 
 test: unittest
 
-test-verbose: install_dependency testee/edge.js
+test-verbose: install_dependency testee/chrome.js testee/edge.js
 	npx mocha test
 
 clean:
